ai-support-assistant-ollama/requirements.txt
fastapi
uvicorn
python-dotenv
requests
numpy
pytest
